feat: refactor components to use shared_components package for improved organization
