+++ TEC-SIS/src/com/fot/system/view/dashboard/admin/components/NoticeFeedPanel.java
@@ -2,7 +2,7 @@ package com.fot.system.view.dashboard.admin.components;
 
 import com.fot.system.config.AppTheme;
 import com.fot.system.model.entity.*;
-import com.fot.system.view.components.FeedItemCard;
+import com.fot.system.view.shared_components.FeedItemCard;
 
 import javax.swing.*;
 import javax.swing.border.EmptyBorder;
