+++ TEC-SIS/src/com/fot/system/view/dashboard/admin/manageCourses/CourseDetailsPanel.java
@@ -3,9 +3,8 @@ package com.fot.system.view.dashboard.admin.manageCourses;
 import com.fot.system.config.AppConfig;
 import com.fot.system.config.AppTheme;
 import com.fot.system.controller.EditCourseController;
-import com.fot.system.model.dto.*;
 import com.fot.system.model.entity.*;
-import com.fot.system.view.components.CustomButton;
+import com.fot.system.view.shared_components.CustomButton;
 import org.kordamp.ikonli.fontawesome5.FontAwesomeSolid;
 import org.kordamp.ikonli.swing.FontIcon;
