+++ TEC-SIS/src/com/fot/system/view/dashboard/admin/manageCourses/ManageCoursesPanel.java
@@ -5,7 +5,7 @@ import com.fot.system.config.AppTheme;
 import com.fot.system.model.dto.*;
 import com.fot.system.model.entity.*;
 import com.fot.system.service.CourseService;
-import com.fot.system.view.components.CustomButton;
+import com.fot.system.view.shared_components.CustomButton;
 import org.kordamp.ikonli.fontawesome5.FontAwesomeSolid;
 import org.kordamp.ikonli.swing.FontIcon;
