+++ TEC-SIS/src/com/fot/system/view/dashboard/admin/manageNotices/AddNewNoticePanel.java
@@ -2,11 +2,10 @@ package com.fot.system.view.dashboard.admin.manageNotices;
 
 import com.fot.system.config.AppTheme;
 import com.fot.system.model.dto.*;
-import com.fot.system.model.entity.*;
-import com.fot.system.view.components.CustomButton;
-import com.fot.system.view.components.ThemedComboBox;
-import com.fot.system.view.components.ThemedDatePicker;
-import com.fot.system.view.components.ThemedTextField;
+import com.fot.system.view.shared_components.CustomButton;
+import com.fot.system.view.shared_components.ThemedComboBox;
+import com.fot.system.view.shared_components.ThemedDatePicker;
+import com.fot.system.view.shared_components.ThemedTextField;
 
 import javax.swing.*;
 import java.awt.*;
