+++ TEC-SIS/src/com/fot/system/view/dashboard/admin/manageUsers/AddNewUserPanel.java
@@ -4,10 +4,10 @@ import com.fot.system.config.AppConfig;
 import com.fot.system.config.AppTheme;
 import com.fot.system.model.dto.*;
 import com.fot.system.model.entity.*;
-import com.fot.system.view.components.CustomButton;
-import com.fot.system.view.components.ProfilePhotoFrame;
-import com.fot.system.view.components.ThemedComboBox;
-import com.fot.system.view.components.ThemedDatePicker;
+import com.fot.system.view.shared_components.CustomButton;
+import com.fot.system.view.shared_components.ProfilePhotoFrame;
+import com.fot.system.view.shared_components.ThemedComboBox;
+import com.fot.system.view.shared_components.ThemedDatePicker;
 
 import javax.swing.*;
 import javax.swing.filechooser.FileNameExtensionFilter;
