+++ TEC-SIS/src/com/fot/system/view/dashboard/admin/manageUsers/UserDetailsPanel.java
@@ -2,10 +2,9 @@ package com.fot.system.view.dashboard.admin.manageUsers;
 
 import com.fot.system.config.AppTheme;
 import com.fot.system.controller.EditUserController;
-import com.fot.system.model.dto.*;
 import com.fot.system.model.entity.*;
-import com.fot.system.view.components.CustomButton;
-import com.fot.system.view.components.ProfilePhotoFrame;
+import com.fot.system.view.shared_components.CustomButton;
+import com.fot.system.view.shared_components.ProfilePhotoFrame;
 import org.kordamp.ikonli.fontawesome5.FontAwesomeSolid;
 import org.kordamp.ikonli.swing.FontIcon;
