+++ TEC-SIS/src/com/fot/system/view/dashboard/lecturer/attendance/AttendanceSessionDialog.java
@@ -3,8 +3,8 @@ package com.fot.system.view.dashboard.lecturer.attendance;
 import com.fot.system.config.AppTheme;
 import com.fot.system.model.dto.*;
 import com.fot.system.model.entity.*;
-import com.fot.system.view.components.CustomButton;
-import com.fot.system.view.components.ThemedDatePicker;
+import com.fot.system.view.shared_components.CustomButton;
+import com.fot.system.view.shared_components.ThemedDatePicker;
 
 import javax.swing.*;
 import javax.swing.border.EmptyBorder;
