+++ TEC-SIS/src/com/fot/system/view/dashboard/lecturer/shared_components/LecturerCourseCard.java
@@ -1,8 +1,7 @@
-package com.fot.system.view.dashboard.lecturer.myCourses;
+package com.fot.system.view.dashboard.lecturer.shared_components;
 
-import com.fot.system.model.dto.*;
 import com.fot.system.model.entity.*;
-import com.fot.system.view.components.CourseSummaryCard;
+import com.fot.system.view.shared_components.CourseSummaryCard;
 
 public class LecturerCourseCard extends CourseSummaryCard {
