+++ TEC-SIS/src/com/fot/system/view/dashboard/lecturer/timetable/TimetablePanel.java
@@ -6,7 +6,7 @@ import com.fot.system.model.dto.*;
 import com.fot.system.model.entity.*;
 import com.fot.system.service.CourseService;
 import com.fot.system.service.TimetableService;
-import com.fot.system.view.components.CustomButton;
+import com.fot.system.view.shared_components.CustomButton;
 import org.kordamp.ikonli.fontawesome5.FontAwesomeSolid;
 import org.kordamp.ikonli.swing.FontIcon;
