+++ TEC-SIS/src/com/fot/system/view/dashboard/student/StudentHomePanel.java
@@ -6,8 +6,8 @@ import com.fot.system.service.CourseMaterialService;
 import com.fot.system.service.CourseService;
 import com.fot.system.service.FileOpenService;
 import com.fot.system.service.NoticeService;
-import com.fot.system.view.components.CloseActionButton;
-import com.fot.system.view.components.CourseSummaryCard;
+import com.fot.system.view.shared_components.CloseActionButton;
+import com.fot.system.view.shared_components.CourseSummaryCard;
 import com.fot.system.view.dashboard.admin.components.NoticeFeedPanel;
 import com.fot.system.view.dashboard.lecturer.myCourses.CourseMaterialCard;
