+++ TEC-SIS/src/com/fot/system/view/dashboard/student/profile/StudentProfilePanel.java
@@ -1,7 +1,7 @@
 package com.fot.system.view.dashboard.student.profile;
 
 import com.fot.system.model.entity.*;
-import com.fot.system.view.dashboard.shared_components.UserProfilePanel;
+import com.fot.system.view.shared_components.UserProfilePanel;
 
 public class StudentProfilePanel extends UserProfilePanel {
