+++ TEC-SIS/src/com/fot/system/view/dashboard/to/TOAttendancePanel.java
@@ -7,7 +7,7 @@ import com.fot.system.model.entity.*;
 import com.fot.system.service.AttendanceService;
 import com.fot.system.service.CourseService;
 import com.fot.system.service.TimetableService;
-import com.fot.system.view.components.CustomButton;
+import com.fot.system.view.shared_components.CustomButton;
 import com.fot.system.view.dashboard.lecturer.attendance.AttendanceSessionDialog;
 import org.kordamp.ikonli.fontawesome5.FontAwesomeSolid;
 import org.kordamp.ikonli.swing.FontIcon;
