+++ TEC-SIS/src/com/fot/system/view/dashboard/to/TOTimetablePanel.java
@@ -2,7 +2,7 @@ package com.fot.system.view.dashboard.to;
 
 import com.fot.system.model.entity.TimetableSession;
 import com.fot.system.model.entity.User;
-import com.fot.system.view.dashboard.shared_components.TimeTablePanel;
+import com.fot.system.view.shared_components.TimeTablePanel;
 
 /**
  * Shows the technical officer timetable using the shared read-only timetable layout.
