+++ TEC-SIS/src/com/fot/system/view/shared_components/StaffProfilePanel.java
@@ -1,4 +1,4 @@
-package com.fot.system.view.dashboard.shared_components;
+package com.fot.system.view.shared_components;
 
 import com.fot.system.model.entity.*;
