+++ TEC-SIS/src/com/fot/system/view/shared_components/ThemedRadioButton.java
@@ -1,4 +1,4 @@
-package com.fot.system.view.components;
+package com.fot.system.view.shared_components;
 
 import com.fot.system.config.AppTheme;
