+++ TEC-SIS/src/com/fot/system/view/shared_components/UserProfilePanel.java
@@ -1,4 +1,4 @@
-package com.fot.system.view.dashboard.shared_components;
+package com.fot.system.view.shared_components;
 
 import com.fot.system.config.AppTheme;
 import com.fot.system.controller.EditUserController;
@@ -6,7 +6,6 @@ import com.fot.system.model.dto.*;
 import com.fot.system.model.entity.*;
 import com.fot.system.service.DepartmentService;
 import com.fot.system.service.UserService;
-import com.fot.system.view.components.CustomButton;
 
 import javax.swing.*;
 import javax.swing.border.EmptyBorder;
