+++ TEC-SIS/src/com/fot/system/view/shared_components/UserProfileViewPanel.java
@@ -1,10 +1,7 @@
-package com.fot.system.view.dashboard.shared_components;
+package com.fot.system.view.shared_components;
 
 import com.fot.system.config.AppTheme;
-import com.fot.system.model.dto.*;
 import com.fot.system.model.entity.*;
-import com.fot.system.view.components.ProfilePhotoFrame;
-import com.fot.system.view.components.ProfileSectionCard;
 
 import javax.swing.*;
 import javax.swing.border.EmptyBorder;
